Add release candidate and remove old implementation
