+++ features/android-studio-updates/src/main/kotlin/features/updates/androidstudio/updatesource/AndroidStudioBlogUpdateSource.kt
@@ -15,38 +15,59 @@
  */
 package features.updates.androidstudio.updatesource
 
-import features.updates.androidstudio.updatesource.rssfetcher.FeedCouldNotBeObtainedException
 import features.updates.androidstudio.updatesource.rssfetcher.Fetcher
 import features.updates.androidstudio.updatesource.rssfetcher.FetcherFactory
-import kotlinx.datetime.Instant
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.merge
 import kotlinx.datetime.toKotlinInstant
 
-/**
- * An implementation of [AndroidStudioUpdateSource] that retrieves updates from the Android Studio Blog.
- */
 internal class AndroidStudioBlogUpdateSource(
     private val source: Fetcher = FetcherFactory.create(),
 ) : AndroidStudioUpdateSource {
 
-    override suspend fun getUpdatesAfter(after: Instant): Result<List<AndroidStudioUpdate>> {
-        return try {
-            val events = source.obtainFeed("https://androidstudio.googleblog.com/feeds/posts/default").entries
-                .stream()
-                .filter { it.publishedOn.toInstant().toKotlinInstant() > after }
-                .map {
-                    AndroidStudioUpdate(
-                        version = extractVersionFromTitle(it.title),
-                        summary = it.title, // TODO Can we extract a better summary?
-                        timestamp = it.publishedOn.toInstant().toKotlinInstant(),
-                        url = it.links.last().url,
-                        updateChannel = extractChannelFromTitle(it.title),
-                    )
-                }
-                .sorted { o1, o2 -> o1.timestamp.compareTo(o2.timestamp) }
-            Result.success(events.toList())
-        } catch (e: FeedCouldNotBeObtainedException) {
-            Result.failure(e)
-        }
+    private val _latestStable = MutableStateFlow<AndroidStudioUpdate?>(null)
+    private val _latestRc = MutableStateFlow<AndroidStudioUpdate?>(null)
+    private val _latestBeta = MutableStateFlow<AndroidStudioUpdate?>(null)
+    private val _latestCanary = MutableStateFlow<AndroidStudioUpdate?>(null)
+
+    override val latestStableUpdate: StateFlow<AndroidStudioUpdate?> = _latestStable
+    override val latestReleaseCandidateUpdate: StateFlow<AndroidStudioUpdate?> = _latestRc
+    override val latestBetaUpdate: StateFlow<AndroidStudioUpdate?> = _latestBeta
+    override val latestCanaryUpdate: StateFlow<AndroidStudioUpdate?> = _latestCanary
+    override val latestUpdate: Flow<AndroidStudioUpdate> =
+        merge(
+            latestStableUpdate.filterNotNull(),
+            latestReleaseCandidateUpdate.filterNotNull(),
+            latestBetaUpdate.filterNotNull(),
+            latestCanaryUpdate.filterNotNull(),
+        )
+
+    override suspend fun checkForUpdates() {
+        val feed = source.obtainFeed("https://androidstudio.googleblog.com/feeds/posts/default")
+        val versionsInFeed = feed.entries
+            .map {
+                AndroidStudioUpdate(
+                    version = extractVersionFromTitle(it.title),
+                    summary = it.title, // TODO Can we extract a better summary?
+                    timestamp = it.publishedOn.toInstant().toKotlinInstant(),
+                    url = it.links.last().url,
+                    updateChannel = extractChannelFromTitle(it.title),
+                )
+            }
+            .sortedByDescending { it.timestamp }
+
+        val lastStableInFeed = versionsInFeed.first { it.updateChannel == AndroidStudioUpdate.UpdateChannel.Stable }
+        val lastRcInFeed = versionsInFeed.first { it.updateChannel == AndroidStudioUpdate.UpdateChannel.ReleaseCandidate }
+        val lastBetaInFeed = versionsInFeed.first { it.updateChannel == AndroidStudioUpdate.UpdateChannel.Beta }
+        val lastCanaryInFeed = versionsInFeed.first { it.updateChannel == AndroidStudioUpdate.UpdateChannel.Canary }
+
+        _latestStable.value = lastStableInFeed
+        _latestRc.value = lastRcInFeed
+        _latestBeta.value = lastBetaInFeed
+        _latestCanary.value = lastCanaryInFeed
     }
 
     internal fun extractVersionFromTitle(title: String): String {
@@ -62,6 +83,7 @@ internal class AndroidStudioBlogUpdateSource(
         return when {
             title.contains("canary", ignoreCase = true) -> AndroidStudioUpdate.UpdateChannel.Canary
             title.contains("beta", ignoreCase = true) -> AndroidStudioUpdate.UpdateChannel.Beta
+            title.contains(" rc ", ignoreCase = true) -> AndroidStudioUpdate.UpdateChannel.ReleaseCandidate
             else -> AndroidStudioUpdate.UpdateChannel.Stable
         }
     }
