+++ features/android-studio-updates/src/main/kotlin/features/updates/androidstudio/updatesource/AndroidStudioUpdate.kt
@@ -39,6 +39,7 @@ data class AndroidStudioUpdate(
      */
     enum class UpdateChannel {
         Stable,
+        ReleaseCandidate,
         Beta,
         Canary
     }
